Split list editor into components and containers
